Don't spawn new threads before main
The function haveSse42 is called at static initialization time. When LOG is
called it will call RAMCloud::Logger::get which will instantiate a static
RAMCloud::Logger instance which spawns a thread
(RAMCloud::Logger::printThreadMain). This thread is spawned before we have a
chance to block signals in it, so the logging thread will intercept signals we
want to handle with our own signal handling thread in
stateless::kernel::signal_handler. To avoid printThreadMain starting before main
we remove the lines in haveSse42 that invoke LOG.