Temp Hack: loopback RoCE with infrc
InfRc transport working with loopback.
This will only allow you to send RoCE messages to yourself.
To send messages remotely we will need to be able to get the GRH information from the remote `QueuePairTuple` object.  This is not currently available so I will need to modify `QueuePairTuple` as well as whatever serialization mechanism is being used to exchange the `qpt` value over UDP in the handshake.