BUG FIX: fix divide by zero error
For arcane reasons, RAMCloud needs to run a disk benchmark before it is capable of maintaining a backup (even in memory).  The basic issue is that it governs its backup “sync to disc” timing based on how fast the disc in question actually is.  However, the in memory backup never calls this benchmark method because there is no disc in question to call it on.  The effect of skipping the benchmark method is to create a repeated divide by zero error in the backup sync logic (due to the fact that the expectedReadMBytesPerSecz value is zero by default).  For some reason the arithmetic exception is caught and the loop restarts, essentially burning an entire CPU core in an endless (and un-logged) error loop.